+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/cores/eventloop/commands/Command.java
@@ -1,0 +1,3 @@
+package org.firstinspires.ftc.teamcode.cores.eventloop.commands;
+
+public interface Command {}
